Restructure packaging for bcmc
COAND-497
