+++ bcmc/src/main/java/com/adyen/checkout/bcmc/internal/ui/BcmcDelegate.kt
@@ -6,8 +6,11 @@
  * Created by oscars on 8/7/2022.
  */
 
-package com.adyen.checkout.bcmc
+package com.adyen.checkout.bcmc.internal.ui
 
+import com.adyen.checkout.bcmc.internal.ui.model.BcmcComponentParams
+import com.adyen.checkout.bcmc.internal.ui.model.BcmcInputData
+import com.adyen.checkout.bcmc.internal.ui.model.BcmcOutputData
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
 import com.adyen.checkout.components.model.payments.request.CardPaymentMethod
@@ -17,7 +20,7 @@ import com.adyen.checkout.components.ui.ViewProvidingDelegate
 import com.adyen.checkout.core.exception.CheckoutException
 import kotlinx.coroutines.flow.Flow
 
-interface BcmcDelegate :
+internal interface BcmcDelegate :
     PaymentComponentDelegate<PaymentComponentState<CardPaymentMethod>>,
     ViewProvidingDelegate,
     ButtonDelegate,
