+++ bcmc/src/main/java/com/adyen/checkout/bcmc/internal/ui/model/BcmcOutputData.kt
@@ -1,17 +1,17 @@
 /*
- * Copyright (c) 2020 Adyen N.V.
+ * Copyright (c) 2023 Adyen N.V.
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by caiof on 27/8/2020.
+ * Created by oscars on 15/2/2023.
  */
-package com.adyen.checkout.bcmc
+package com.adyen.checkout.bcmc.internal.ui.model
 
 import com.adyen.checkout.card.internal.ui.model.ExpiryDate
 import com.adyen.checkout.components.base.OutputData
 import com.adyen.checkout.components.ui.FieldState
 
-data class BcmcOutputData internal constructor(
+internal data class BcmcOutputData internal constructor(
     val cardNumberField: FieldState<String>,
     val expiryDateField: FieldState<ExpiryDate>,
     val cardHolderNameField: FieldState<String>,
